Remove spring dependency
