--- apm-protocol/apm-ui-protocol/README.md
@@ -1,6 +1,0 @@
-## Using graphiql 
-
-Graphiql */ˈɡrafək(ə)l/* is interactive in-browser GraphQL IDE which can be used by following steps.
-
- 1. Run `org.apache.skywalking.apm.ui.protocol.GraphQLInitializer#main`
- 1. Access [http://localhost:8080/graphiql](http://localhost:8080/graphiql)
